check_split_validity function to annotate nhood_summary
A check was performed in order to decide whether or not a cluster
should be split. A cluster was considered a valid split if it had 100
cells or 2.5% of all cells (arbitrary values, could discuss and change).
If a cluster failed these checks, it was merged with its nearest cluster
until all clusters are pass the check or only one cluster remains.
In the NFOL example we end up having only NFOL.pos because the
first condition (>2.5% cells) is met but the second (>=100 cells) is not